Created debugging utils file. Moved functions.
